Feat: home 화면 서버페이지로 변경 useQuery가 아닌 next13에서 업데이트 된 getStaticProps와 비슷한 동작을  하는 fetch 함수로 변경
